+++ dill/src/utils/graphviz.rs
@@ -3,17 +3,17 @@ use std::fmt::Write;
 use crate::*;
 
 fn get_type_name(i: &TypeInfo) -> String {
-    let iang = i.type_name.find('<').unwrap_or(i.type_name.len());
-    let icol = i.type_name[0..iang].rfind("::").map(|i| i + 2).unwrap_or(0);
+    let iang = i.name.find('<').unwrap_or(i.name.len());
+    let icol = i.name[0..iang].rfind("::").map(|i| i + 2).unwrap_or(0);
 
-    format!("\"{}\"", &i.type_name[icol..iang])
+    format!("\"{}\"", &i.name[icol..iang])
 }
 
 fn get_spec_name(i: &DependencyInfo) -> String {
     let spec = i
         .spec
-        .type_name
-        .replace(i.type_info.type_name, "")
+        .name
+        .replace(i.iface.name, "")
         .replace("dill::specs::", "");
 
     match spec.as_str() {
@@ -55,16 +55,16 @@ pub fn render(cat: &Catalog) -> String {
     .unwrap();
 
     let mut builders: Vec<_> = cat.builders().collect();
-    builders.sort_by_key(|b| b.instance_type().type_name);
+    builders.sort_by_key(|b| b.instance_type().name);
 
     for b in &builders {
         let inst = b.instance_type();
 
         let mut ifaces = b.interfaces_get_all();
-        ifaces.sort_by_key(|i| i.type_name);
+        ifaces.sort_by_key(|i| i.name);
 
         let mut deps = b.dependencies_get_all();
-        deps.sort_by_key(|i| i.type_info.type_name);
+        deps.sort_by_key(|i| i.iface.name);
 
         for iface in &ifaces {
             writeln!(
@@ -81,7 +81,7 @@ pub fn render(cat: &Catalog) -> String {
                 s,
                 "    {} -> {} [label=\"{}\", arrowhead=vee]",
                 get_type_name(&inst),
-                get_type_name(&dep.type_info),
+                get_type_name(&dep.iface),
                 get_spec_name(dep)
             )
             .unwrap();